修改一个强转导致的编译错误

+++ libasm/snprintf.cc
@@ -383,7 +383,7 @@ static size_t dopr(char *buffer, size_t maxlen, const char *format, va_list args
 						break;
 					case 'p':
 						strvalue = va_arg (args, char *);
-						fmtint (buffer, &currlen, maxlen, (long) strvalue, 16, min, max, flags);
+						fmtint (buffer, &currlen, maxlen, (long)(unsigned long long) strvalue, 16, min, max, flags);
 						break;
 					case 'n':
 						if (cflags == DP_C_SHORT) {
